Add files via upload
In this case study, we first converted a color picture to grayscale image. After that, motion blur was added to the gray image. The intensity of blur was controlled by a variable and the angle was taken at random. Speckle noise was added to the blurred image. Using Inverse Filter and Wiener Filter, both motion blurred and noisy motion blurred images are restored and both of them are compared.